ちょっと修正

+++ roles/ubuntu/default.rb
@@ -22,10 +22,10 @@ package 'x11-xserver-utils'
 package 'fonts-ipafont'
 execute 'locale-gen ja_JP.UTF-8'
 
-execute 'remove libpulse0' do
+execute 'remove libpulse0'  do
   command <<-EOL
     apt purge -y libpulse0
   EOL
 
-  if 'uname -a | grep Microsoft'
+  only_if 'uname -a | grep Microsoft'
 end
